+++ app/controllers/product/card_folios_controller.ts
@@ -4,10 +4,9 @@ import { errors as lucidErrors } from '@adonisjs/lucid'
 import { errors as vineErrors } from '@vinejs/vine'
 import NotFoundException from '#exceptions/not_found_exception'
 import FolioService from '#services/folio_service'
-import { SuccessOutputDto } from '#types/success_output_dto_type'
+import { SuccessOutputDTO } from '#types/success_output_dto_type'
 import {
   collectValidator,
-  createFolioValidator,
   occurrenceValidator,
   removeMainValidator,
 } from '#validators/card_folio_validator'
@@ -15,8 +14,6 @@ import CardService from '#services/card_service'
 import CardFolioService from '#services/card_folio_service'
 import ValidationException from '#exceptions/validation_exception'
 import DuplicateEntryException from '#exceptions/duplicate_entry_exception'
-import InsufficientCardOccurrenceException from '#exceptions/card_occurrence_exception'
-import CardNotOwnedException from '#exceptions/card_not_owned_exception'
 
 @inject()
 export default class CardFoliosController {
@@ -32,7 +29,7 @@ export default class CardFoliosController {
       const card = await this.cardService.getCardIdById(payload.cardId)
       const folio = await this.folioService.getMainFolioByUserId(authUser.id)
       await this.cardFolioService.createCardMainFolio(card.id, folio.id)
-      const successResponse: SuccessOutputDto = {
+      const successResponse: SuccessOutputDTO = {
         message: 'Card added to your main folio successfully',
       }
       return response.ok(successResponse)
@@ -56,7 +53,7 @@ export default class CardFoliosController {
       const card = await this.cardService.getCardIdById(payload.params.id)
       const folio = await this.folioService.getMainFolioByUserId(authUser.id)
       await this.cardFolioService.updateCardFolioOccurrence(card.id, folio.id, payload.occurrence)
-      const successResponse: SuccessOutputDto = {
+      const successResponse: SuccessOutputDTO = {
         message: 'Card occurrence updated successfully',
       }
       return response.ok(successResponse)
@@ -77,53 +74,8 @@ export default class CardFoliosController {
       const card = await this.cardService.getCardIdById(payload.params.id)
       const folio = await this.folioService.getMainFolioByUserId(authUser.id)
       await this.cardFolioService.deleteCardFromFolioByCardIdAndFolioId(card.id, folio.id)
-      const successResponse: SuccessOutputDto = {
+      const successResponse: SuccessOutputDTO = {
         message: 'Card remove from main folio successfully',
-      }
-      return response.ok(successResponse)
-    } catch (error) {
-      if (error instanceof vineErrors.E_VALIDATION_ERROR) {
-        throw new ValidationException(error)
-      }
-      if (error instanceof lucidErrors.E_ROW_NOT_FOUND) {
-        throw new NotFoundException(error)
-      }
-      throw error
-    }
-  }
-
-  async store({ request, response, authUser }: HttpContext) {
-    try {
-      const payload = await createFolioValidator.validate(request.all())
-      const mainFolio = await this.folioService.getMainFolioByUserId(authUser.id)
-
-      for (const card of payload.cards) {
-        await this.cardService.getCardIdById(card.id)
-        const cardInMainFolio = await this.cardFolioService.getCardFolioByCardIdAndFolioId(
-          card.id,
-          mainFolio.id
-        )
-
-        if (!cardInMainFolio) {
-          throw new CardNotOwnedException(card.id)
-        }
-
-        if (card.occurrence > cardInMainFolio.occurrence) {
-          throw new InsufficientCardOccurrenceException(
-            card.id,
-            card.occurrence,
-            cardInMainFolio.occurrence
-          )
-        }
-      }
-
-      const folio = await this.folioService.createFolio(authUser.id, payload.name, payload.imageUrl)
-      for (const card of payload.cards) {
-        await this.cardFolioService.createCardFolio(card.id, folio.id, card.occurrence)
-      }
-
-      const successResponse: SuccessOutputDto = {
-        message: 'Folio created successfully',
       }
       return response.ok(successResponse)
     } catch (error) {
