+++ app/controllers/product/folios_controller.ts
@@ -3,7 +3,7 @@ import { inject } from '@adonisjs/core'
 import { errors as lucidErrors } from '@adonisjs/lucid'
 import NotFoundException from '#exceptions/not_found_exception'
 import FolioService from '#services/folio_service'
-import { SuccessOutputDto } from '#types/success_output_dto_type'
+import { SuccessOutputDTO } from '#types/success_output_dto_type'
 
 @inject()
 export default class FoliosController {
@@ -12,7 +12,7 @@ export default class FoliosController {
   async init({ response, authUser }: HttpContext) {
     try {
       await this.folioService.createMainFolio(authUser.id)
-      const successResponse: SuccessOutputDto = {
+      const successResponse: SuccessOutputDTO = {
         message: 'Folio initialized successfully',
       }
       return response.ok(successResponse)
