+++ app/types/success_output_dto_type.ts
@@ -1,3 +1,3 @@
-export interface SuccessOutputDto {
+export interface SuccessOutputDTO {
   message: string
 }
