+++ app/validators/card_folio_validator.ts
@@ -25,16 +25,3 @@ export const removeMainValidator = vine.compile(
     }),
   })
 )
-
-export const createFolioValidator = vine.compile(
-  vine.object({
-    imageUrl: vine.string(),
-    name: vine.string().minLength(1).maxLength(100),
-    cards: vine.array(
-      vine.object({
-        id: vine.string(),
-        occurrence: vine.number().min(1).max(1000),
-      })
-    ),
-  })
-)
