+++ start/routes.ts
@@ -35,7 +35,7 @@ router
             router.get('/statistics', [MainFoliosController, 'statistics']).use(middleware.auth())
             router.get('/:id', [ChildFoliosController, 'show']).use(middleware.auth())
             router.get('/:id/cards', [ChildFoliosController, 'cards']).use(middleware.auth())
-            router.post('/', [CardFoliosController, 'store']).use(middleware.auth())
+            router.post('/', [ChildFoliosController, 'store']).use(middleware.auth())
             router.post('/init', [FoliosController, 'init']).use(middleware.auth())
             router.post('/cards', [CardFoliosController, 'collect']).use(middleware.auth())
             router.patch('/cards/:id', [CardFoliosController, 'occurrence']).use(middleware.auth())
